Organização em mais de um arquivo. Mais legível e organizado. Não corrigi o problema de previsão do commit anterior.
